Remove unused queue and worker pool implementations
