+++ pool/runner_test.go
@@ -36,15 +36,8 @@ func makeTestQueue(pool amboy.Runner) amboy.Queue {
 
 func TestRunnerImplementations(t *testing.T) {
 	pools := map[string]func() amboy.Runner{
-		"Local":  func() amboy.Runner { return new(localWorkers) },
-		"Single": func() amboy.Runner { return new(single) },
-		"Noop":   func() amboy.Runner { return new(noopPool) },
-		"RateLimitedSimple": func() amboy.Runner {
-			return &simpleRateLimited{
-				size:     1,
-				interval: time.Second,
-			}
-		},
+		"Local": func() amboy.Runner { return new(localWorkers) },
+		"Noop":  func() amboy.Runner { return new(noopPool) },
 		"RateLimitedAverage": func() amboy.Runner {
 			return &ewmaRateLimiting{
 				size:   1,
