+++ queue/group_remote_mongo.go
@@ -2,11 +2,7 @@ package queue
 
 import (
 	"context"
-	"fmt"
 	"regexp"
-	"runtime"
-	"strings"
-	"sync"
 	"time"
 
 	"github.com/mongodb/amboy"
@@ -16,17 +12,6 @@ import (
 	"github.com/pkg/errors"
 	"go.mongodb.org/mongo-driver/bson"
 )
-
-// remoteMongoQueueGroup is a group of queues backed by MongoDB where each queue
-// belongs to its own collection for namespace isolation.
-type remoteMongoQueueGroup struct {
-	collPrefix string
-	canceler   context.CancelFunc
-	mu         sync.RWMutex
-	opts       MongoDBQueueGroupOptions
-	queues     map[string]amboy.Queue
-	ttlMap     map[string]time.Time
-}
 
 // MongoDBQueueGroupOptions describe options to create a queue group backed by
 // MongoDB.
@@ -106,61 +91,54 @@ func (o MongoDBQueueGroupOptions) getQueueOptsWithPrecedence(id string, opts ...
 	return mergeMongoDBQueueOptions(precedenceOrderedOpts...)
 }
 
-type listCollectionsOutput struct {
-	Name string `bson:"name"`
+// remoteMongoQueueGroupSingle is a group of queues backed by MongoDB that are
+// multiplexed into a single MongoDB collection and isolated by group
+// namespaces.
+type remoteMongoQueueGroupSingle struct {
+	canceler context.CancelFunc
+	opts     MongoDBQueueGroupOptions
+	cache    GroupCache
 }
 
-// NewMongoDBQueueGroup constructs a new remote queue group with the given
-// prefix for its collection names. If the TTL is 0, the queues will not be
-// TTLed except when the client explicitly calls Prune.
+// NewMongoDBSingleQueueGroup constructs a new remote queue group. If the TTL is
+// 0, the queues will not be TTLed except when the client explicitly calls
+// Prune.
 //
-// The MongoDB remote queue group creates a new collection for every queue. This
-// is probably most viable for lower volume workloads; however, the caching
-// mechanism may be more responsive in some situations.
-func NewMongoDBQueueGroup(ctx context.Context, collPrefix string, opts MongoDBQueueGroupOptions) (amboy.QueueGroup, error) {
+// The MongoDB single remote queue group multiplexes all queues into a single
+// collection.
+func NewMongoDBSingleQueueGroup(ctx context.Context, opts MongoDBQueueGroupOptions) (amboy.QueueGroup, error) {
 	if opts.DefaultQueue.DB == nil {
 		return nil, errors.New("must provide DB options")
 	}
 
-	// Because of the way this queue is implemented, the driver's collection
-	// name has group information, but the jobs within the collection don't have
-	// any group information. Therefore, we have to set UseGroups to false so
-	// that the driver will treat the jobs as if they're in their own queue.
-	opts.DefaultQueue.DB.UseGroups = false
-	opts.DefaultQueue.DB.GroupName = ""
-
-	// Collection must be provided for queue options, but the collection's name
-	// is not yet known here; the collection's name is determine when the queue
-	// is generated dynamically in Get. Therefore, the Collection set here is
-	// not actually important and is only necessary to pass validation; the
-	// actual collection name will be validated when the queue is created.
-	originalCollName := opts.DefaultQueue.DB.Collection
-	opts.DefaultQueue.DB.Collection = "placeholder"
+	// Because of the way this queue group is implemented, the driver must
+	// account for job isolation between multiplexed queues within a single
+	// collection, so it needs to set UseGroups.
+	opts.DefaultQueue.DB.UseGroups = true
+	// GroupName must be provided if UseGroups is set, but the queue's name is
+	// not yet known here; the queue's name is determined when the queue is
+	// generated dynamically in Get. Therefore, the GroupName set here is not
+	// actually important and is only necessary to pass validation; the actual
+	// queue's name will be validated when the queue is created.
+	originalGroupName := opts.DefaultQueue.DB.GroupName
+	opts.DefaultQueue.DB.GroupName = "placeholder"
 
 	if err := opts.validate(); err != nil {
 		return nil, errors.Wrap(err, "invalid remote queue options")
 	}
 
-	opts.DefaultQueue.DB.Collection = originalCollName
+	opts.DefaultQueue.DB.GroupName = originalGroupName
 
 	ctx, cancel := context.WithCancel(ctx)
-	g := &remoteMongoQueueGroup{
-		collPrefix: collPrefix,
-		canceler:   cancel,
-		opts:       opts,
-		queues:     map[string]amboy.Queue{},
-		ttlMap:     map[string]time.Time{},
-	}
-
-	if opts.PruneFrequency > 0 && opts.TTL > 0 {
-		if err := g.Prune(ctx); err != nil {
-			return nil, errors.Wrap(err, "pruning queue")
-		}
+	g := &remoteMongoQueueGroupSingle{
+		canceler: cancel,
+		opts:     opts,
+		cache:    NewGroupCache(opts.TTL),
 	}
 
 	if opts.PruneFrequency > 0 && opts.TTL > 0 {
 		go func() {
-			defer recovery.LogStackTraceAndContinue("panic in remote queue group ticker")
+			defer recovery.LogStackTraceAndContinue("remote queue group background prune")
 			ticker := time.NewTicker(opts.PruneFrequency)
 			defer ticker.Stop()
 			for {
@@ -168,7 +146,7 @@ func NewMongoDBQueueGroup(ctx context.Context, collPrefix string, opts MongoDBQu
 				case <-ctx.Done():
 					return
 				case <-ticker.C:
-					grip.Error(message.WrapError(g.Prune(ctx), "pruning remote queue group"))
+					grip.Error(message.WrapError(g.Prune(ctx), "pruning remote queue group database"))
 				}
 			}
 		}()
@@ -176,15 +154,15 @@ func NewMongoDBQueueGroup(ctx context.Context, collPrefix string, opts MongoDBQu
 
 	if opts.BackgroundCreateFrequency > 0 {
 		go func() {
-			defer recovery.LogStackTraceAndContinue("panic in remote queue group ticker")
-			ticker := time.NewTicker(opts.PruneFrequency)
+			defer recovery.LogStackTraceAndContinue("remote queue group background queue creation")
+			ticker := time.NewTicker(opts.BackgroundCreateFrequency)
 			defer ticker.Stop()
 			for {
 				select {
 				case <-ctx.Done():
 					return
 				case <-ticker.C:
-					grip.Error(message.WrapError(g.startQueues(ctx), "starting queues"))
+					grip.Error(message.WrapError(g.startQueues(ctx), "starting external queues"))
 				}
 			}
 		}()
@@ -197,98 +175,80 @@ func NewMongoDBQueueGroup(ctx context.Context, collPrefix string, opts MongoDBQu
 	return g, nil
 }
 
-func (g *remoteMongoQueueGroup) startQueues(ctx context.Context) error {
-	g.mu.Lock()
-	defer g.mu.Unlock()
+// getQueues return all queues that are active. A queue is active if it still
+// has jobs to process (still has pending, in progress, or retrying jobs), or if
+// it recently completed a job within the TTL. Inactive queues (i.e. queues that
+// have no jobs to process and have not recently completed a job within the
+// TTL) are not returned.
+func (g *remoteMongoQueueGroupSingle) getQueues(ctx context.Context) ([]string, error) {
+	cursor, err := g.opts.DefaultQueue.DB.Client.Database(g.opts.DefaultQueue.DB.DB).Collection(addGroupSuffix(g.opts.DefaultQueue.DB.Collection)).Aggregate(ctx,
+		[]bson.M{
+			{
+				"$match": bson.M{
+					"$or": []bson.M{
+						{
+							"status.completed": false,
+						},
+						{
+							"status.completed": true,
+							"status.mod_ts":    bson.M{"$gte": time.Now().Add(-g.opts.TTL)},
+						},
+						{
+							"status.completed":       true,
+							"retry_info.retryable":   true,
+							"retry_info.needs_retry": true,
+						},
+					},
+				},
+			},
+			{
 
-	colls, err := g.getExistingCollections(ctx)
+				"$group": bson.M{
+					"_id": nil,
+					"groups": bson.M{
+						"$addToSet": "$group",
+					},
+				},
+			},
+		},
+	)
 	if err != nil {
-		return errors.Wrap(err, "getting existing collections")
+		return nil, errors.WithStack(err)
+	}
+
+	out := struct {
+		Groups []string `bson:"groups"`
+	}{}
+
+	catcher := grip.NewBasicCatcher()
+	for cursor.Next(ctx) {
+		if err = cursor.Decode(&out); err != nil {
+			catcher.Add(err)
+		} else {
+			break
+		}
+	}
+	catcher.Add(cursor.Err())
+	catcher.Add(cursor.Close(ctx))
+
+	return out.Groups, catcher.Resolve()
+}
+
+func (g *remoteMongoQueueGroupSingle) startQueues(ctx context.Context) error {
+	queues, err := g.getQueues(ctx)
+	if err != nil {
+		return errors.WithStack(err)
 	}
 
 	catcher := grip.NewBasicCatcher()
-	for _, coll := range colls {
-		q, err := g.startProcessingRemoteQueue(ctx, coll)
-		if err != nil {
-			catcher.Add(errors.Wrap(err, "starting queue"))
-		} else {
-			g.queues[g.idFromCollection(coll)] = q
-			g.ttlMap[g.idFromCollection(coll)] = time.Now()
-		}
+	// Refresh the TTLs on all the queues that were recently accessed or still
+	// have jobs to run.
+	for _, id := range queues {
+		_, err := g.Get(ctx, id)
+		catcher.Add(err)
 	}
 
 	return catcher.Resolve()
-}
-
-func (g *remoteMongoQueueGroup) Len() int {
-	g.mu.RLock()
-	defer g.mu.RUnlock()
-
-	return len(g.queues)
-}
-
-func (g *remoteMongoQueueGroup) Queues(ctx context.Context) []string {
-	g.mu.RLock()
-	defer g.mu.RUnlock()
-
-	out, _ := g.getExistingCollections(ctx)
-
-	return out
-}
-
-func (g *remoteMongoQueueGroup) startProcessingRemoteQueue(ctx context.Context, coll string, opts ...amboy.QueueOptions) (amboy.Queue, error) {
-	id := g.idFromCollection(coll)
-	// The driver already adds the jobs suffix implicitly to the collection
-	// name.
-	coll = trimJobsSuffix(coll)
-	mdbOpts, err := getMongoDBQueueOptions(opts...)
-	if err != nil {
-		return nil, errors.Wrap(err, "getting queue options")
-	}
-
-	queueOpts := g.opts.getQueueOptsWithPrecedence(id, mdbOpts...)
-
-	// The collection name has to be set to ensure the queue uses its
-	// collection-level namespace.
-	queueOpts.DB.Collection = coll
-	// These settings must apply to all queues in the queue group because if the
-	// queue-specific options differ from the defaults, it will affect the queue
-	// group's ability to manage jobs properly.
-	queueOpts.DB.UseGroups = g.opts.DefaultQueue.DB.UseGroups
-	queueOpts.DB.GroupName = g.opts.DefaultQueue.DB.GroupName
-
-	q, err := queueOpts.buildQueue(ctx)
-	if err != nil {
-		return nil, errors.Wrap(err, "constructing queue")
-	}
-
-	if err := q.Start(ctx); err != nil {
-		return nil, errors.Wrap(err, "starting queue")
-	}
-	return q, nil
-}
-
-func (g *remoteMongoQueueGroup) getExistingCollections(ctx context.Context) ([]string, error) {
-	c, err := g.opts.DefaultQueue.DB.Client.Database(g.opts.DefaultQueue.DB.DB).ListCollections(ctx, bson.M{"name": bson.M{"$regex": fmt.Sprintf("^%s.*", g.collPrefix)}})
-	if err != nil {
-		return nil, errors.Wrap(err, "calling listCollections")
-	}
-	defer c.Close(ctx)
-	var collections []string
-	for c.Next(ctx) {
-		elem := listCollectionsOutput{}
-		if err := c.Decode(&elem); err != nil {
-			return nil, errors.Wrap(err, "parsing listCollections output")
-		}
-		collections = append(collections, elem.Name)
-	}
-	if err := c.Err(); err != nil {
-		return nil, errors.Wrap(err, "iterating over list collections cursor")
-	}
-	if err := c.Close(ctx); err != nil {
-		return nil, errors.Wrap(err, "closing cursor")
-	}
-	return collections, nil
 }
 
 // Get a queue with the given id. Get sets the last accessed time to now. Note
@@ -296,165 +256,74 @@ func (g *remoteMongoQueueGroup) getExistingCollections(ctx context.Context) ([]s
 // the caller actually performs an operation on the queue (e.g. add a job) must
 // be within the TTL; otherwise, the queue might be closed before the operation
 // is done.
-func (g *remoteMongoQueueGroup) Get(ctx context.Context, id string, opts ...amboy.QueueOptions) (amboy.Queue, error) {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	if queue, ok := g.queues[id]; ok {
-		g.ttlMap[id] = time.Now()
-		return queue, nil
+func (g *remoteMongoQueueGroupSingle) Get(ctx context.Context, id string, opts ...amboy.QueueOptions) (amboy.Queue, error) {
+	var queue remoteQueue
+	var err error
+
+	switch q := g.cache.Get(id).(type) {
+	case remoteQueue:
+		return q, nil
+	case nil:
+		mdbOpts, err := getMongoDBQueueOptions(opts...)
+		if err != nil {
+			return nil, errors.Wrap(err, "getting queue options")
+		}
+
+		queueOpts := g.opts.getQueueOptsWithPrecedence(id, mdbOpts...)
+
+		// The group name has to be set to ensure the queue uses its namespace
+		// within the single multiplexed collection.
+		queueOpts.DB.GroupName = id
+		// These settings must apply to all queues in the queue group to ensure
+		// proper management of jobs within the single multiplexed collection.
+		queueOpts.DB.UseGroups = g.opts.DefaultQueue.DB.UseGroups
+		queueOpts.DB.Collection = g.opts.DefaultQueue.DB.Collection
+
+		queue, err = queueOpts.buildQueue(ctx)
+		if err != nil {
+			return nil, errors.Wrap(err, "constructing queue")
+		}
+	default:
+		return q, nil
 	}
 
-	queue, err := g.startProcessingRemoteQueue(ctx, g.collectionFromID(id), opts...)
-	if err != nil {
+	if err = g.cache.Set(id, queue, g.opts.TTL); err != nil {
+		// If another thread already created and set the queue in the cache in
+		// between the cache check above and now, it should be safe to throw
+		// away the queue that was just constructed since it hasn't started yet.
+		queue.Close(ctx)
+
+		if q := g.cache.Get(id); q != nil {
+			return q, nil
+		}
+
+		return nil, errors.Wrap(err, "caching queue")
+	}
+
+	if err := queue.Start(ctx); err != nil {
 		return nil, errors.Wrap(err, "starting queue")
 	}
-	g.queues[id] = queue
-	g.ttlMap[id] = time.Now()
+
 	return queue, nil
 }
 
-// Put a queue with the given ID. The caller is responsible for starting the
-// queue.
-func (g *remoteMongoQueueGroup) Put(ctx context.Context, id string, queue amboy.Queue) error {
-	g.mu.RLock()
-	if _, ok := g.queues[id]; ok {
-		g.mu.RUnlock()
-		return errors.New("a queue already exists at this index")
-	}
-	g.mu.RUnlock()
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	// Check again in case the map was modified after we released the read lock.
-	if _, ok := g.queues[id]; ok {
-		return errors.New("a queue already exists at this index")
-	}
-
-	g.queues[id] = queue
-	g.ttlMap[id] = time.Now()
-	return nil
+func (g *remoteMongoQueueGroupSingle) Put(ctx context.Context, name string, queue amboy.Queue) error {
+	return g.cache.Set(name, queue, 0)
 }
 
-// Prune queues that have no pending work, and have completed work older than the TTL.
-func (g *remoteMongoQueueGroup) Prune(ctx context.Context) error {
-	g.mu.Lock()
-	defer g.mu.Unlock()
+func (g *remoteMongoQueueGroupSingle) Len() int { return g.cache.Len() }
 
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	colls, err := g.getExistingCollections(ctx)
-	if err != nil {
-		return errors.Wrap(err, "getting collections")
-	}
-	collsToCheck := []string{}
-	for _, coll := range colls {
-		// This is an optimization. If we've added to the queue recently enough, there's no
-		// need to query its contents, since it cannot be old enough to prune.
-		if t, ok := g.ttlMap[g.idFromCollection(coll)]; !ok || ok && time.Since(t) > g.opts.TTL {
-			collsToCheck = append(collsToCheck, coll)
-		}
-	}
-	catcher := grip.NewBasicCatcher()
-	wg := &sync.WaitGroup{}
-	collsDeleteChan := make(chan string, len(collsToCheck))
-	collsDropChan := make(chan string, len(collsToCheck))
-
-	for _, coll := range collsToCheck {
-		collsDropChan <- coll
-	}
-	close(collsDropChan)
-
-	wg = &sync.WaitGroup{}
-	for i := 0; i < runtime.NumCPU(); i++ {
-		wg.Add(1)
-		go func(ch chan string) {
-			defer recovery.LogStackTraceAndContinue("panic in pruning collections")
-			defer wg.Done()
-			for nextColl := range collsDropChan {
-				c := g.opts.DefaultQueue.DB.Client.Database(g.opts.DefaultQueue.DB.DB).Collection(nextColl)
-				count, err := c.CountDocuments(ctx, bson.M{
-					"status.completed": true,
-					"status.in_prog":   false,
-					"$or": []bson.M{
-						{"status.mod_ts": bson.M{"$gte": time.Now().Add(-g.opts.TTL)}},
-						{
-							"retry_info.retryable":   true,
-							"retry_info.needs_retry": true,
-						},
-					},
-				})
-				if err != nil {
-					catcher.Add(err)
-					return
-				}
-				if count > 0 {
-					return
-				}
-				count, err = c.CountDocuments(ctx, bson.M{"status.completed": false})
-				if err != nil {
-					catcher.Add(err)
-					return
-				}
-				if count > 0 {
-					return
-				}
-				if queue, ok := g.queues[g.idFromCollection(nextColl)]; ok {
-					queue.Close(ctx)
-					select {
-					case <-ctx.Done():
-						return
-					case ch <- g.idFromCollection(nextColl):
-						// pass
-					}
-				}
-				if err := c.Drop(ctx); err != nil {
-					catcher.Add(err)
-				}
-			}
-		}(collsDeleteChan)
-	}
-	wg.Wait()
-	close(collsDeleteChan)
-	for id := range collsDeleteChan {
-		delete(g.queues, id)
-		delete(g.ttlMap, id)
-	}
-
-	return catcher.Resolve()
+func (g *remoteMongoQueueGroupSingle) Queues(ctx context.Context) []string {
+	queues, err := g.getQueues(ctx)
+	grip.Warning(message.WrapError(err, "getting active queues in queue group"))
+	return queues
 }
 
-// Close the queues.
-func (g *remoteMongoQueueGroup) Close(ctx context.Context) error {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	g.canceler()
-	waitCh := make(chan struct{})
-	wg := &sync.WaitGroup{}
-	go func() {
-		defer recovery.LogStackTraceAndContinue("panic in remote queue group closer")
-		for _, queue := range g.queues {
-			wg.Add(1)
-			go func(queue amboy.Queue) {
-				defer recovery.LogStackTraceAndContinue("panic in remote queue group closer")
-				defer wg.Done()
-				queue.Close(ctx)
-			}(queue)
-		}
-		wg.Wait()
-		close(waitCh)
-	}()
-	select {
-	case <-waitCh:
-		return nil
-	case <-ctx.Done():
-		return errors.WithStack(ctx.Err())
-	}
+func (g *remoteMongoQueueGroupSingle) Prune(ctx context.Context) error {
+	return g.cache.Prune(ctx)
 }
 
-func (g *remoteMongoQueueGroup) collectionFromID(id string) string {
-	return addJobsSuffix(g.collPrefix + id)
-}
-
-func (g *remoteMongoQueueGroup) idFromCollection(collection string) string {
-	return trimJobsSuffix(strings.TrimPrefix(collection, g.collPrefix))
+func (g *remoteMongoQueueGroupSingle) Close(ctx context.Context) error {
+	defer g.canceler()
+	return g.cache.Close(ctx)
 }
