+++ queue/queue_test.go
@@ -51,14 +51,11 @@ type QueueTestCase struct {
 	MaxSize             int
 	SingleWorker        bool
 	MultiSupported      bool
-	OrderedSupported    bool
-	OrderedStartsBefore bool
 	WaitUntilSupported  bool
 	DispatchBySupported bool
 	MaxTimeSupported    bool
 	ScopesSupported     bool
 	RetrySupported      bool
-	SkipUnordered       bool
 	IsRemote            bool
 	Skip                bool
 }
@@ -81,38 +78,6 @@ type SizeTestCase struct {
 func DefaultQueueTestCases() []QueueTestCase {
 	return []QueueTestCase{
 		{
-			Name:                "AdaptiveOrdering",
-			OrderedSupported:    true,
-			OrderedStartsBefore: true,
-			WaitUntilSupported:  true,
-			SingleWorker:        true,
-			DispatchBySupported: true,
-			MaxTimeSupported:    true,
-			MinSize:             2,
-			MaxSize:             16,
-			Constructor: func(ctx context.Context, _ string, size int) (amboy.Queue, TestCloser, error) {
-				return NewAdaptiveOrderedLocalQueue(size, defaultLocalQueueCapcity), func(ctx context.Context) error { return nil }, nil
-			},
-		},
-		{
-			Name:                "LocalOrdered",
-			OrderedStartsBefore: false,
-			OrderedSupported:    true,
-			MinSize:             2,
-			MaxSize:             8,
-			Constructor: func(ctx context.Context, _ string, size int) (amboy.Queue, TestCloser, error) {
-				return NewLocalOrdered(size), func(ctx context.Context) error { return nil }, nil
-			},
-		},
-		{
-			Name:             "Priority",
-			ScopesSupported:  true,
-			MaxTimeSupported: true,
-			Constructor: func(ctx context.Context, _ string, size int) (amboy.Queue, TestCloser, error) {
-				return NewLocalPriorityQueue(size, defaultLocalQueueCapcity), func(ctx context.Context) error { return nil }, nil
-			},
-		},
-		{
 			Name:                "LimitedSize",
 			WaitUntilSupported:  true,
 			DispatchBySupported: true,
@@ -132,25 +97,6 @@ func DefaultQueueTestCases() []QueueTestCase {
 			Constructor: func(ctx context.Context, _ string, size int) (amboy.Queue, TestCloser, error) {
 				q, err := NewLocalLimitedSizeSerializable(size, 1024*size)
 				return q, func(ctx context.Context) error { return nil }, err
-			},
-		},
-		{
-			Name:             "Shuffled",
-			SingleWorker:     true,
-			MaxTimeSupported: true,
-			ScopesSupported:  true,
-			Constructor: func(ctx context.Context, _ string, size int) (amboy.Queue, TestCloser, error) {
-				return NewLocalShuffled(size, defaultLocalQueueCapcity), func(ctx context.Context) error { return nil }, nil
-			},
-		},
-		{
-			Name:    "SQSFifo",
-			MaxSize: 4,
-			Skip:    true,
-			Constructor: func(ctx context.Context, _ string, size int) (amboy.Queue, TestCloser, error) {
-				q, err := NewSQSFifoQueue(randomString(4), size, awsTestCredentialsFromEnv())
-				closer := func(ctx context.Context) error { return nil }
-				return q, closer, err
 			},
 		},
 	}
@@ -176,7 +122,7 @@ func MergeQueueTestCases(ctx context.Context, cases ...[]QueueTestCase) <-chan Q
 func MongoDBQueueTestCases(client *mongo.Client) []QueueTestCase {
 	return []QueueTestCase{
 		{
-			Name:               "MongoUnordered",
+			Name:               "MongoRemote",
 			IsRemote:           true,
 			WaitUntilSupported: true,
 			MaxTimeSupported:   true,
@@ -190,7 +136,6 @@ func MongoDBQueueTestCases(client *mongo.Client) []QueueTestCase {
 				opts := MongoDBQueueOptions{
 					DB:         &mdbOpts,
 					NumWorkers: utility.ToIntPtr(size),
-					Ordered:    utility.FalsePtr(),
 				}
 				q, err := NewMongoDBQueue(ctx, opts)
 				if err != nil {
@@ -217,7 +162,7 @@ func MongoDBQueueTestCases(client *mongo.Client) []QueueTestCase {
 			},
 		},
 		{
-			Name:               "MongoGroupUnordered",
+			Name:               "GroupMongoRemote",
 			IsRemote:           true,
 			WaitUntilSupported: true,
 			MaxTimeSupported:   true,
@@ -233,7 +178,6 @@ func MongoDBQueueTestCases(client *mongo.Client) []QueueTestCase {
 				opts := MongoDBQueueOptions{
 					DB:         &mdbOpts,
 					NumWorkers: utility.ToIntPtr(size),
-					Ordered:    utility.FalsePtr(),
 				}
 				q, err := NewMongoDBQueue(ctx, opts)
 				if err != nil {
@@ -261,7 +205,7 @@ func MongoDBQueueTestCases(client *mongo.Client) []QueueTestCase {
 			},
 		},
 		{
-			Name:               "MongoUnorderedMGOBSON",
+			Name:               "MongoRemoteMGOBSON",
 			IsRemote:           true,
 			WaitUntilSupported: true,
 			MaxTimeSupported:   true,
@@ -276,49 +220,6 @@ func MongoDBQueueTestCases(client *mongo.Client) []QueueTestCase {
 				opts := MongoDBQueueOptions{
 					DB:         &mdbOpts,
 					NumWorkers: utility.ToIntPtr(size),
-					Ordered:    utility.FalsePtr(),
-				}
-				q, err := NewMongoDBQueue(ctx, opts)
-				if err != nil {
-					return nil, nil, err
-				}
-				rq, ok := q.(remoteQueue)
-				if !ok {
-					return nil, nil, errors.New("invalid queue constructed")
-				}
-
-				closer := func(ctx context.Context) error {
-					catcher := grip.NewBasicCatcher()
-					d := rq.Driver()
-					if d != nil {
-						catcher.Add(d.Close(ctx))
-					}
-
-					catcher.Add(client.Database(mdbOpts.DB).Collection(addJobsSuffix(name)).Drop(ctx))
-
-					return catcher.Resolve()
-				}
-
-				return q, closer, nil
-			},
-		},
-		{
-			Name:               "MongoOrdered",
-			IsRemote:           true,
-			WaitUntilSupported: true,
-			MaxTimeSupported:   true,
-			RetrySupported:     true,
-			ScopesSupported:    true,
-			OrderedSupported:   true,
-			Constructor: func(ctx context.Context, name string, size int) (amboy.Queue, TestCloser, error) {
-				mdbOpts := defaultMongoDBTestOptions()
-				mdbOpts.Client = client
-				mdbOpts.Collection = name
-				mdbOpts.Format = amboy.BSON2
-				opts := MongoDBQueueOptions{
-					DB:         &mdbOpts,
-					NumWorkers: utility.ToIntPtr(size),
-					Ordered:    utility.TruePtr(),
 				}
 				q, err := NewMongoDBQueue(ctx, opts)
 				if err != nil {
@@ -345,7 +246,6 @@ func MongoDBQueueTestCases(client *mongo.Client) []QueueTestCase {
 			},
 		},
 	}
-
 }
 
 func DefaultPoolTestCases() []PoolTestCase {
@@ -355,37 +255,9 @@ func DefaultPoolTestCases() []PoolTestCase {
 			SetPool: func(q amboy.Queue, _ int) error { return nil },
 		},
 		{
-			Name:      "Single",
-			SkipMulti: true,
-			MinSize:   1,
-			MaxSize:   1,
-			SetPool: func(q amboy.Queue, _ int) error {
-				runner := pool.NewSingle()
-				if err := runner.SetQueue(q); err != nil {
-					return err
-				}
-
-				return q.SetRunner(runner)
-			},
-		},
-		{
 			Name:    "Abortable",
 			MinSize: 4,
 			SetPool: func(q amboy.Queue, size int) error { return q.SetRunner(pool.NewAbortablePool(size, q)) },
-		},
-		{
-			Name:         "RateLimitedSimple",
-			MinSize:      4,
-			MaxSize:      16,
-			RateLimiting: true,
-			SetPool: func(q amboy.Queue, size int) error {
-				runner, err := pool.NewSimpleRateLimitedWorkers(size, 10*time.Millisecond, q)
-				if err != nil {
-					return nil
-				}
-
-				return q.SetRunner(runner)
-			},
 		},
 		{
 			Name:         "RateLimitedAverage",
@@ -471,16 +343,10 @@ func TestQueueSmoke(t *testing.T) {
 						}
 
 						t.Run("MaxSize"+size.Name, func(t *testing.T) {
-							if !test.SkipUnordered {
-								t.Run("Unordered", func(t *testing.T) {
-									unorderedTest(bctx, t, test, runner, size)
-								})
-							}
-							if test.OrderedSupported {
-								t.Run("Ordered", func(t *testing.T) {
-									orderedTest(bctx, t, test, runner, size)
-								})
-							}
+							t.Run("Unordered", func(t *testing.T) {
+								unorderedTest(bctx, t, test, runner, size)
+							})
+
 							if test.WaitUntilSupported {
 								testWaitUntilOnce.Do(func() {
 									t.Run("WaitUntil", func(t *testing.T) {
@@ -517,7 +383,7 @@ func TestQueueSmoke(t *testing.T) {
 							})
 
 							if test.ScopesSupported {
-								if test.SingleWorker && (!test.OrderedSupported || test.OrderedStartsBefore) && size.Size >= 4 {
+								if test.SingleWorker && size.Size >= 4 {
 									testScopesOnce.Do(func() {
 										t.Run("ScopedLock", func(t *testing.T) {
 											scopedLockTest(bctx, t, test, runner, size)
@@ -546,9 +412,6 @@ func TestQueueSmoke(t *testing.T) {
 							}
 
 							t.Run("SaveLockingCheck", func(t *testing.T) {
-								if test.OrderedSupported && !test.OrderedStartsBefore {
-									t.Skip("test does not support queues where queues don't accept work after dispatching")
-								}
 								ctx, cancel := context.WithCancel(bctx)
 								defer cancel()
 								name := newDriverID()
@@ -609,12 +472,6 @@ func unorderedTest(bctx context.Context, t *testing.T, test QueueTestCase, runne
 	defer func() { require.NoError(t, closer(ctx)) }()
 	require.NoError(t, runner.SetPool(q, size.Size))
 
-	if test.OrderedSupported && !test.OrderedStartsBefore {
-		// pass
-	} else {
-		require.NoError(t, q.Start(ctx))
-	}
-
 	testNames := []string{"test", "second", "workers", "forty-two", "true", "false", ""}
 	numJobs := size.Size * len(testNames)
 
@@ -635,9 +492,8 @@ func unorderedTest(bctx context.Context, t *testing.T, test QueueTestCase, runne
 		}(i)
 	}
 	wg.Wait()
-	if test.OrderedSupported && !test.OrderedStartsBefore {
-		require.NoError(t, q.Start(ctx))
-	}
+
+	require.NoError(t, q.Start(ctx))
 
 	amboy.WaitInterval(ctx, q, 100*time.Millisecond)
 
@@ -686,9 +542,6 @@ func orderedTest(bctx context.Context, t *testing.T, test QueueTestCase, runner 
 	require.NoError(t, err)
 	defer os.RemoveAll(tempDir)
 
-	if test.OrderedStartsBefore {
-		require.NoError(t, q.Start(ctx))
-	}
 	for i := 0; i < size.Size/2; i++ {
 		for _, name := range testNames {
 			fn := filepath.Join(tempDir, fmt.Sprintf("%s.%d", name, i))
@@ -703,9 +556,7 @@ func orderedTest(bctx context.Context, t *testing.T, test QueueTestCase, runner 
 		}
 	}
 
-	if !test.OrderedStartsBefore {
-		require.NoError(t, q.Start(ctx))
-	}
+	require.NoError(t, q.Start(ctx))
 
 	require.Equal(t, numJobs, q.Stats(ctx).Total, fmt.Sprintf("with %d workers", size.Size))
 	amboy.WaitInterval(ctx, q, 50*time.Millisecond)
@@ -914,10 +765,6 @@ func oneExecutionTest(bctx context.Context, t *testing.T, test QueueTestCase, ru
 	mockJobCounters.Reset()
 	count := 40
 
-	if !test.OrderedSupported || test.OrderedStartsBefore {
-		require.NoError(t, q.Start(ctx))
-	}
-
 	for i := 0; i < count; i++ {
 		j := newMockJob()
 		jobID := fmt.Sprintf("%d.%d.mock.single-exec", i, job.GetNumber())
@@ -925,9 +772,7 @@ func oneExecutionTest(bctx context.Context, t *testing.T, test QueueTestCase, ru
 		assert.NoError(t, q.Put(ctx, j))
 	}
 
-	if test.OrderedSupported && !test.OrderedStartsBefore {
-		require.NoError(t, q.Start(ctx))
-	}
+	require.NoError(t, q.Start(ctx))
 
 	amboy.WaitInterval(ctx, q, 100*time.Millisecond)
 	assert.Equal(t, count, mockJobCounters.Count())
